Rename and move shared components
