+++ site/pages/our-work/case-study/bank/index.js
@@ -22,12 +22,12 @@ import {
   SectionBody,
   SectionHeading,
   UL,
-} from '../case-study-components';
+} from '../shared';
 
 import Card from '../../../../components/card';
 import ListBox from '../../../../components/list-box';
 
-import WhatToReadNext from '../../what-to-read-next';
+import WhatToReadNext from '../shared/what-to-read-next';
 import ChecklistContactUs from '../../../../slices/checklist-contact-us-slice';
 
 import ScrollTracker from '../../../../components/scroll-tracker';
