+++ site/pages/our-work/case-study/car-trawler/index.js
@@ -11,8 +11,8 @@ import {
   SectionHeading,
   SectionBody,
   P,
-} from '../case-study-components';
-import WhatToReadNext from '../../what-to-read-next';
+} from '../shared';
+import WhatToReadNext from '../shared/what-to-read-next';
 import ChecklistContactUs from '../../../../slices/checklist-contact-us-slice';
 import styles from './style.css';
